Reorganize Parser unit tests
